Add kmp-common-network module
