+++ appJs/appApiJs/build.gradle.kts
@@ -6,11 +6,14 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
+                // TODO: These should not be directly accessed
+                implementation(projects.kmpCommonFirebase)
+                implementation(projects.kmpCommonNetwork)
+                implementation(projects.moduleA)
+                // Needed to get the Firebase deployment to work
                 implementation(npm("firebase", "9.10.0"))
                 implementation(npm("firebase-admin", "11.0.1"))
                 implementation(npm("firebase-functions", "3.24.0"))
-                implementation(projects.kmpCommonFirebase)
-                implementation(projects.moduleA)
             }
         }
     }
