+++ appJs/appApiJs/src/commonMain/kotlin/com/gchristov/thecodinglove/Main.kt
@@ -1,19 +1,12 @@
 package com.gchristov.thecodinglove
 
 import com.gchristov.thecodinglove.kmpcommonfirebase.CommonFirebaseModule
-import dev.gitlive.firebase.Firebase
-import dev.gitlive.firebase.FirebaseApp
-import dev.gitlive.firebase.firestore.firestore
-import io.ktor.client.*
+import com.gchristov.thecodinglove.kmpcommonfirebase.CommonNetworkModule
 import io.ktor.client.call.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.client.plugins.logging.*
 import io.ktor.client.request.*
-import io.ktor.serialization.kotlinx.json.*
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
 
 external fun require(module:String) : dynamic
 external var exports: dynamic
@@ -21,7 +14,7 @@ external var exports: dynamic
 fun main(args: Array<String>) {
     val fireFunctions = require("firebase-functions")
     exports.myTestFun = fireFunctions.https.onRequest { request, response ->
-        val client = provideHttpClient()
+        val client = CommonNetworkModule.injectHttpClient()
         GlobalScope.launch {
             val userResponse: Response = client.get("https://reqres.in/api/users").body()
 
@@ -36,20 +29,6 @@ fun main(args: Array<String>) {
 
             response.send(Messenger().message() + ", " + userResponse.page + ", " + count)
         }
-    }
-}
-
-private fun provideHttpClient() = HttpClient {
-    install(ContentNegotiation) {
-        json(
-            Json {
-                ignoreUnknownKeys = true
-            }
-        )
-    }
-    install(Logging) {
-        logger = Logger.SIMPLE
-        level = LogLevel.ALL
     }
 }
