+++ kmp/kmp-gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/js/JavascriptModulePlugin.kt
@@ -1,6 +1,5 @@
 package com.gchristov.thecodinglove.gradleplugins.js
 
-import com.gchristov.thecodinglove.gradleplugins.Deps
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
@@ -33,21 +32,13 @@ class JavascriptBrowserExecutablePlugin : JavascriptBrowserTargetPlugin() {
 open class JavascriptNodeModulePlugin : JavascriptNodeTargetPlugin() {
     override fun apply(target: Project) {
         super.apply(target)
-        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            sourceSets.maybeCreate("commonMain").dependencies {
-                implementation(kotlin("test"))
-                implementation(project(":kmp-common-di"))
-            }
-        }
         // TODO: Move this to a better location
         target.plugins.apply("org.jetbrains.kotlin.plugin.serialization")
         target.extensions.configure(KotlinMultiplatformExtension::class.java) {
             sourceSets.maybeCreate("commonMain").dependencies {
-                implementation(Deps.Ktor.client)
-                implementation(Deps.Ktor.contentNegotiation)
-                implementation(Deps.Ktor.serialisation)
-                implementation(Deps.Ktor.logging)
-                implementation(Deps.Ktor.logback)
+                // TODO: Move this to a better location
+                implementation(kotlin("test"))
+                implementation(project(":kmp-common-di"))
             }
         }
     }
