+++ modules.gradle.kts
@@ -5,6 +5,7 @@
 val projects = listOf(
     "kmp-common-di",
     "kmp-common-firebase",
+    "kmp-common-network",
     "kmp-gradle-plugins",
     "kmp-module-b",
     "appApiJs",
